Refactor utility imports across components to improve code consistency
- Updated import paths for the `cn` utility function from `@/lib/utils` to `@/utils/utils` in multiple components, ensuring a unified import structure.
- Removed the obsolete `utils.ts` file to streamline the codebase and enhance maintainability.
