+++ src/components/option-lab/option-type-select.tsx
@@ -3,7 +3,7 @@ import { FormControl, FormItem, FormLabel, FormMessage } from "@/components/ui/f
 import { Button } from "@/components/ui/button";
 import { useFormContext } from 'react-hook-form';
 import type { FormData } from '@/types/mint/form';
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 
 export const OptionTypeSelector = () => {
   const { setValue, watch } = useFormContext<FormData>();
