+++ src/components/trade/option-chain-table.tsx
@@ -8,7 +8,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 import {
   Tooltip,
   TooltipContent,
