+++ src/components/ui/dropdown-menu.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 import { Check, ChevronRight, Circle } from "lucide-react"
 
 const DropdownMenu = DropdownMenuPrimitive.Root
